Corrected numerous errors and now controller compiles
Corrected numerous cyclic includes
Corrected Math class override cmath/math.h
Removed standard namespace calls
Corrected abstract feedback controller instantiation
Allowed HMatrix class to take doubles as inputs for clean calls - casts to int
Corrected dividing by zero in kalman filter
Instantiated quadcopter class and tested position and rotation calculation
Created Motor, Servo, and MPU6050 with simple setup
Moved Quaternion creation from two vectors to rotation class

+++ DTRQController/HMatrix.h
@@ -2,7 +2,7 @@
 
 #include <EulerAngles.h>
 #include <EulerOrder.h>
-#include <Math.h>
+#include <Mathematics.h>
 
 typedef struct HMatrix {
 public:
@@ -12,5 +12,5 @@ public:
 	HMatrix(std::vector<std::vector<double>> hMatrix);
 	HMatrix(int x, int y) : hierarchicalMatrix(x, std::vector<double>(y, double())) {}
 
-	double& operator() (int x, int y) { return hierarchicalMatrix[x][y]; }
+	double& operator() (double x, double y) { return hierarchicalMatrix[(int)x][(int)y]; }
 } HMatrix;
